requirements.txt
pandas==2.2.3
numpy==2.1.1
tqdm==4.66.5
torch
torchvision==0.16.0     #Should match the version of torch
requests==2.32.3
transformers==4.45.2
Pillow==10.4.0

